+++ cedar-policy/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added an API to `ParseError` to quickly get the primary source span
 - Added an API, `unknown_entities`, to `PolicySet` to collect unknown entity UIDs from `PartialResponse`.
 - Added APIs `remove`, `remove_template` and `unlink` to remove policies from the `PolicySet`
+- Added API `get_linked_policies` to get the policies linked to a `Template`
 
 ### Changed